requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.0
gradio==4.8.0
chromadb==0.4.15
sentence-transformers==2.2.2
pypdf2==3.0.1
python-dotenv==1.0.0
pydantic==2.4.2
pytest==7.4.3
requests==2.31.0
torch==2.1.0+cu118
torchvision==0.16.0+cu118
torchaudio==2.1.0+cu118
transformers==4.35.0
accelerate==0.24.1
numpy==1.24.3
regex==2023.10.3
